Move CreateUser class to a differrent file. Include create_student method. Include create_teacher method.
